Extract all_ems/all_valid_ems/any_valid? methods
These methods were essentially duplicated in the PerEmsWorkerMixin and
PerEmsTypeWorkerMixin, as well has being required by provider workers
that aren't PerEms or PerEmsType (e.g. the old VimBroker and the current
AmazonCoordinator)
