+++ app/models/manageiq/providers/base_manager/event_catcher.rb
@@ -1,6 +1,7 @@
 class ManageIQ::Providers::BaseManager::EventCatcher < MiqWorker
   require_nested :Runner
 
+  include ProviderWorkerMixin
   include PerEmsWorkerMixin
 
   self.required_roles = ["event"]
