+++ app/models/manageiq/providers/base_manager/metrics_collector_worker.rb
@@ -3,6 +3,7 @@ class ManageIQ::Providers::BaseManager::MetricsCollectorWorker < MiqQueueWorkerB
 
   require_nested :Runner
 
+  include ProviderWorkerMixin
   include PerEmsTypeWorkerMixin
 
   self.required_roles = ["ems_metrics_collector"]
