+++ app/models/manageiq/providers/base_manager/operations_worker.rb
@@ -1,6 +1,7 @@
 class ManageIQ::Providers::BaseManager::OperationsWorker < MiqQueueWorkerBase
   require_nested :Runner
 
+  include ProviderWorkerMixin
   include PerEmsWorkerMixin
 
   self.required_roles = %w[ems_operations]
