+++ app/models/mixins/per_ems_type_worker_mixin.rb
@@ -7,17 +7,5 @@ module PerEmsTypeWorkerMixin
 
       super
     end
-
-    def ems_class
-      module_parent
-    end
-
-    def emses_in_zone
-      ems_class.where(:zone_id => MiqServer.my_server.zone.id)
-    end
-
-    def any_valid_ems_in_zone?
-      emses_in_zone.any?(&:authentication_status_ok?)
-    end
   end
 end
