+++ app/models/mixins/per_ems_worker_mixin.rb
@@ -10,18 +10,6 @@ module PerEmsWorkerMixin
   end
 
   module ClassMethods
-    def ems_class
-      module_parent
-    end
-
-    def all_ems_in_zone
-      ems_class.where(:zone_id => MiqServer.my_server.zone.id)
-    end
-
-    def all_valid_ems_in_zone
-      all_ems_in_zone.select { |e| e.enabled && e.authentication_status_ok? }
-    end
-
     def desired_queue_names
       all_valid_ems_in_zone.collect { |e| queue_name_for_ems(e) }
     end
